DRY'd Rot3D kernels for unit tests
