+++ Makefile
@@ -63,6 +63,8 @@ OBJ =                         \
   VideoCore/Invoke.o          \
   VideoCore/VideoCore.o
 
+LIB = $(patsubst %,$(OBJ_DIR)/%,$(OBJ))
+
 
 # All programs in the Examples directory
 EXAMPLES =  \
@@ -79,14 +81,26 @@ EXAMPLES =  \
 	HeatMap
 
 EXAMPLE_TARGETS = $(patsubst %,$(OBJ_DIR)/bin/%,$(EXAMPLES))
-LIB = $(patsubst %,$(OBJ_DIR)/%,$(OBJ))
 
-# List of dependencies defined from list of object files
-# Note that the example programs in Examples are not included here
+
+# Example object files
+EXAMPLES_EXTRA = \
+	Rot3DKernels.o
+
+EXAMPLES_OBJ = $(patsubst %,$(OBJ_DIR)/Examples/%,$(EXAMPLES_EXTRA))
+#$(info $(EXAMPLES_OBJ))
+
+
+# Dependencies from list of object files
 DEPS := $(LIB:.o=.d)
 #$(info $(DEPS))
-
 -include $(DEPS)
+
+# Dependencies for the include files in the Examples directory.
+# Basically, every .h file under examples has a .d in the build directory
+EXAMPLES_DEPS = $(EXAMPLES_OBJ:.o=.d)
+#$(info $(EXAMPLES_DEPS))
+-include $(EXAMPLES_DEPS)
 
 
 # Top-level targets
@@ -147,6 +161,7 @@ $(OBJ_DIR)/%.o: $(ROOT)/%.cpp | $(OBJ_DIR)
 #
 # Targets for Examples
 #
+$(OBJ_DIR)/bin/Rot3D: $(OBJ_DIR)/Examples/Rot3DKernels.o
 
 $(OBJ_DIR)/bin/%: $(OBJ_DIR)/Examples/%.o $(QPU_LIB)
 	@echo Linking $@...
@@ -155,6 +170,8 @@ $(OBJ_DIR)/bin/%: $(OBJ_DIR)/Examples/%.o $(QPU_LIB)
 $(OBJ_DIR)/Examples/%.o: Examples/%.cpp | $(OBJ_DIR)
 	@echo Compiling $<
 	@$(CXX) -c $(CXX_FLAGS) -o $@ $<
+
+
 
 $(EXAMPLES) :% :$(OBJ_DIR)/bin/%
 
@@ -178,7 +195,7 @@ UNIT_TESTS =          \
 
 # For some reason, doing an interim step to .o results in linkage errors (undefined references).
 # So this target compiles the source files directly to the executable.
-$(OBJ_DIR)/bin/runTests: $(UNIT_TESTS) | $(QPU_LIB)
+$(OBJ_DIR)/bin/runTests: $(UNIT_TESTS) $(EXAMPLES_OBJ) | $(QPU_LIB)
 	@$(CXX) $(CXX_FLAGS) $^ -L$(OBJ_DIR) -lQPULib -o $@
 
 test : all $(OBJ_DIR)/bin/runTests
