+++ Tests/testRot3D.cpp
@@ -1,89 +1,6 @@
 #include "catch.hpp"
-#include <QPULib.h>
 #include <math.h>
-
-//
-// The following is taken from the Rot3D example, for the purpose of illustrating
-// the unit tests. 
-// TODO: DRY this code and the example.
-//
-
-// ============================================================================
-// Scalar version
-// ============================================================================
-
-void rot3D(int n, float cosTheta, float sinTheta, float* x, float* y)
-{
-  for (int i = 0; i < n; i++) {
-    float xOld = x[i];
-    float yOld = y[i];
-    x[i] = xOld * cosTheta - yOld * sinTheta;
-    y[i] = yOld * cosTheta + xOld * sinTheta;
-  }
-}
-
-// ============================================================================
-// Vector version 1
-// ============================================================================
-
-void rot3D_1(Int n, Float cosTheta, Float sinTheta, Ptr<Float> x, Ptr<Float> y)
-{
-  For (Int i = 0, i < n, i = i+16)
-    Float xOld = x[i];
-    Float yOld = y[i];
-    x[i] = xOld * cosTheta - yOld * sinTheta;
-    y[i] = yOld * cosTheta + xOld * sinTheta;
-  End
-}
-
-// ============================================================================
-// Vector version 2
-// ============================================================================
-
-void rot3D_2(Int n, Float cosTheta, Float sinTheta, Ptr<Float> x, Ptr<Float> y)
-{
-  Int inc = 16;
-  Ptr<Float> p = x + index();
-  Ptr<Float> q = y + index();
-  gather(p); gather(q);
- 
-  Float xOld, yOld;
-  For (Int i = 0, i < n, i = i+inc)
-    gather(p+inc); gather(q+inc); 
-    receive(xOld); receive(yOld);
-    store(xOld * cosTheta - yOld * sinTheta, p);
-    store(yOld * cosTheta + xOld * sinTheta, q);
-    p = p+inc; q = q+inc;
-  End
-
-  receive(xOld); receive(yOld);
-}
-
-
-// ============================================================================
-// Vector version 3
-// ============================================================================
-
-void rot3D_3(Int n, Float cosTheta, Float sinTheta, Ptr<Float> x, Ptr<Float> y)
-{
-  Int inc = numQPUs() << 4;
-  Ptr<Float> p = x + index() + (me() << 4);
-  Ptr<Float> q = y + index() + (me() << 4);
-  gather(p); gather(q);
- 
-  Float xOld, yOld;
-  For (Int i = 0, i < n, i = i+inc)
-    gather(p+inc); gather(q+inc); 
-    receive(xOld); receive(yOld);
-    store(xOld * cosTheta - yOld * sinTheta, p);
-    store(yOld * cosTheta + xOld * sinTheta, q);
-    p = p+inc; q = q+inc;
-  End
-
-  receive(xOld); receive(yOld);
-}
-
-
+#include "../Examples/Rot3DKernels.h"
 
 // ============================================================================
 // Support routines
